+++ packages/ui/styles/src/theme/borderWidths.ts
@@ -10,6 +10,7 @@ export type PikasBorderWidths = typeof pikasBorderWidths;
 export type PikasBorderWidth = keyof PikasBorderWidths;
 
 export type BorderWidthsRecordValue = number | string;
+// eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
 export type BorderWidthsRecordKey = PikasBorderWidth | number | string;
 export type BorderWidthsRecord = Record<
   BorderWidthsRecordKey,
@@ -26,4 +27,4 @@ export const loadBorderWidths = <T extends BorderWidthsRecord>(
   ({
     ...pikasBorderWidths,
     ...values,
-  } as PikasBorderWidths & T);
+  }) as PikasBorderWidths & T;