+++ packages/ui/styles/src/theme/transitions.ts
@@ -4,6 +4,7 @@ export type PikasTransitions = typeof pikasTransitions;
 export type PikasTransition = keyof PikasTransitions;
 
 export type TransitionsRecordValue = string;
+// eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
 export type TransitionsRecordKey = PikasTransition | number | string;
 export type TransitionsRecord = Record<
   TransitionsRecordKey,
@@ -20,4 +21,4 @@ export const loadTransitions = <T extends TransitionsRecord>(
   ({
     ...pikasTransitions,
     ...values,
-  } as PikasTransitions & T);
+  }) as PikasTransitions & T;