+++ packages/ui/styles/src/utils/index.ts
@@ -1,8 +1,8 @@
-import type { ConfigType } from '@stitches/react/types/config'; // TODO: remove this line when @stitches/react is updated
+import type { ConfigType } from '@secrecy/stitches-react/types/config';
 
 export const loadUtils = <T extends Record<string, string>>(
   utils: T
 ): ConfigType.Utils<T> =>
   ({
     ...utils,
-  } as ConfigType.Utils<T>);
+  }) as ConfigType.Utils<T>;