+++ packages/ui/avatar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/avatar
 
+## 4.0.0
+
+### Major Changes
+
+- a7e6f0b: upgrade deps
+
+### Patch Changes
+
+- Updated dependencies [a7e6f0b]
+  - @pikas-ui/skeleton@4.0.0
+  - @pikas-ui/styles@5.0.0
+
 ## 3.1.0
 
 ### Minor Changes