+++ packages/ui/context-menu/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pikas-ui/context-menu
 
+## 5.0.0
+
+### Major Changes
+
+- a7e6f0b: upgrade deps
+
+### Patch Changes
+
+- Updated dependencies [a7e6f0b]
+  - @pikas-ui/icons@5.0.0
+  - @pikas-ui/loader@5.0.0
+  - @pikas-ui/menu@5.0.0
+  - @pikas-ui/styles@5.0.0
+
 ## 4.2.0
 
 ### Minor Changes