+++ packages/ui/dialog/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pikas-ui/dialog
 
+## 4.0.0
+
+### Major Changes
+
+- a7e6f0b: upgrade deps
+
+### Patch Changes
+
+- Updated dependencies [a7e6f0b]
+  - @pikas-ui/button@5.0.0
+  - @pikas-ui/icons@5.0.0
+  - @pikas-ui/loader@5.0.0
+  - @pikas-ui/slider@5.0.0
+  - @pikas-ui/styles@5.0.0
+  - @pikas-ui/title@3.0.0
+
 ## 3.1.0
 
 ### Minor Changes